Utilization no longer rails-specific, moved to judo-ruby
