+++ judoscale-rails/lib/judoscale/rails.rb
@@ -3,10 +3,9 @@
 require "judoscale-ruby"
 require "judoscale/version"
 require "judoscale/rails/railtie"
-require "judoscale/rails/metrics_collector"
 require "rails/version"
 
 Judoscale.add_adapter :"judoscale-rails", {
   adapter_version: Judoscale::VERSION,
   framework_version: ::Rails.version
-}, metrics_collector: Judoscale::Rails::MetricsCollector
+}, metrics_collector: Judoscale::WebMetricsCollector
