+++ judoscale-rails/lib/judoscale/rails/railtie.rb
@@ -4,7 +4,7 @@ require "rails"
 require "rails/engine"
 require "judoscale/request_middleware"
 require "judoscale/rails/config"
-require "judoscale/rails/utilization_middleware"
+require "judoscale/utilization_middleware"
 require "judoscale/logger"
 require "judoscale/reporter"
