--- judoscale-rails/lib/judoscale/rails/utilization_middleware.rb
@@ -1,109 +1,0 @@
-# frozen_string_literal: true
-
-require "singleton"
-
-module Judoscale
-  module Rails
-    class UtilizationMiddleware
-      def initialize(app)
-        @app = app
-      end
-
-      def call(env)
-        tracker = UtilizationTracker.instance
-        tracker.start! unless tracker.started?
-
-        tracker.incr
-
-        @app.call(env)
-      ensure
-        tracker.decr
-      end
-    end
-
-    class UtilizationTracker
-      include Singleton
-
-      def initialize
-        @active_request_counter = 0
-        @mutex = Mutex.new
-        @started = false
-      end
-
-      def start!
-        @mutex.synchronize do
-          @started = true
-          init_idle_report_cycle!
-        end
-      end
-
-      def started?
-        @started
-      end
-
-      def incr
-        @mutex.synchronize do
-          if @active_request_counter == 0 && @idle_started_at
-            # We were idle and now we're not - add to total idle time
-            @total_idle_time += get_current_time - @idle_started_at
-            @idle_started_at = nil
-          end
-
-          @active_request_counter += 1
-        end
-      end
-
-      def decr
-        @mutex.synchronize do
-          @active_request_counter -= 1
-
-          if @active_request_counter == 0
-            # We're now idle - start tracking idle time
-            @idle_started_at = get_current_time
-          end
-        end
-      end
-
-      def idle_ratio(reset: true)
-        @mutex.synchronize do
-          return 0.0 if @report_cycle_started_at.nil?
-
-          current_time = get_current_time
-
-          total_report_cycle_time = current_time - @report_cycle_started_at
-
-          return 0.0 if total_report_cycle_time <= 0
-
-          # Capture remaining idle time
-          if @idle_started_at
-            @total_idle_time += current_time - @idle_started_at
-            @idle_started_at = current_time
-          end
-
-          idle_ratio = @total_idle_time / total_report_cycle_time
-
-          reset_idle_report_cycle!(current_time: current_time) if reset
-
-          idle_ratio
-        end
-      end
-
-      private
-
-      def get_current_time
-        Process.clock_gettime(Process::CLOCK_MONOTONIC)
-      end
-
-      def init_idle_report_cycle!
-        current_time = get_current_time
-        @idle_started_at = current_time
-        reset_idle_report_cycle!(current_time: current_time)
-      end
-
-      def reset_idle_report_cycle!(current_time:)
-        @total_idle_time = 0.0
-        @report_cycle_started_at = current_time
-      end
-    end
-  end
-end
