+++ judoscale-rails/test/adapter_test.rb
@@ -8,7 +8,7 @@ module Judoscale
     it "adds itself as an adapter with information to be reported to the Judoscale API" do
       adapter = Judoscale.adapters.detect { |adapter| adapter.identifier == :"judoscale-rails" }
       _(adapter).wont_be_nil
-      _(adapter.metrics_collector).must_equal Judoscale::Rails::MetricsCollector
+      _(adapter.metrics_collector).must_equal Judoscale::WebMetricsCollector
 
       report = ::Judoscale::Report.new(Judoscale.adapters, Judoscale::Config.instance, [])
       _(report.as_json[:adapters]).must_include(:"judoscale-rails")
